+++ src/dice/todo.rs
@@ -1,22 +1,14 @@
 use crate::prelude::*;
 
-/// Generator for prototyping that always panics when generating a value.
+/// Generator for prototyping that panics when created.
 ///
 /// The `todo!()` macro doesn't work in places where a `impl Die` or `impl DieOnce` is expected
 /// because the type `!` is not stabilized yet, hence it's not possible to implement `Die` or
-/// `DieOnce` for `!`.
+/// `DieOnce` for `!`. This function can be used as an alternative.
 ///
 /// # Examples
 ///
-/// This example doesn't panic because the generator isn't used:
-///
-/// ```
-/// use dicetest::prelude::*;
-///
-/// let _byte_die = dice::todo::<u8>();
-/// ```
-///
-/// However, this example panics:
+/// This example panics:
 ///
 /// ```should_panic
 /// use dicetest::prelude::*;
@@ -26,13 +18,13 @@ use crate::prelude::*;
 /// let limit = Limit::default();
 /// let mut fate = Fate::new(&mut prng, limit);
 ///
-/// let _byte = fate.roll(dice::todo::<u32>());
+/// let _number_die = dice::todo::<u32>();
 /// ```
 pub fn todo<T>() -> impl Die<T> {
-    dice::from_fn(|_| {
-        panic!(
-            "implementation for Die<{}> is missing",
-            std::any::type_name::<T>(),
-        )
-    })
+    panic!(
+        "implementation for Die<{}> is missing",
+        std::any::type_name::<T>(),
+    );
+    #[allow(unreachable_code)]
+    dice::from_fn(|_| unreachable!())
 }